Combine all the buffer sources into one .h/.cc pair.
