+++ tiledb/sm/buffer/preallocated_buffer.h
@@ -27,89 +27,5 @@
  *
  * @section DESCRIPTION
  *
- * This file defines class PreallocatedBuffer.
+ * This file does nothing.
  */
-
-#ifndef TILEDB_PREALLOCATED_BUFFER_H
-#define TILEDB_PREALLOCATED_BUFFER_H
-
-#include <cinttypes>
-
-#include "tiledb/common/status.h"
-#include "tiledb/sm/buffer/buffer.h"
-
-using namespace tiledb::common;
-
-namespace tiledb {
-namespace sm {
-
-/**
- * Convenience class wrapping a pre-allocated memory region in a class that
- * tracks an offset when reading and writing to the region.
- */
-class PreallocatedBuffer : public BufferBase {
- public:
-  /* ********************************* */
-  /*     CONSTRUCTORS & DESTRUCTORS    */
-  /* ********************************* */
-
-  /**
-   * Constructor.
-   *
-   * @param data The data of the buffer.
-   * @param size The size of the buffer.
-   */
-  PreallocatedBuffer(const void* data, uint64_t size);
-
-  /* ********************************* */
-  /*                API                */
-  /* ********************************* */
-
-  /** Returns the buffer data pointer at the current offset. */
-  void* cur_data() const;
-
-  /** Returns the buffer data. */
-  const void* data() const;
-
-  /** Returns the "free space" in the buffer, which is the size minus the
-   * current offset. */
-  uint64_t free_space() const;
-
-  /**
-   * Returns a value from the buffer of type T.
-   *
-   * @tparam T The type of the value to be read.
-   * @param offset The offset in the local buffer to read from.
-   * @return The desired value of type T.
-   */
-  template <class T>
-  inline T value(uint64_t offset) const {
-    return ((const T*)(((const char*)data_) + offset))[0];
-  }
-
-  /**
-   * Returns the value at the current offset of the buffer of type T.
-   *
-   * @tparam T The type of the value to be returned.
-   * @return The value to be returned of type T.
-   */
-  template <class T>
-  inline T value() const {
-    return ((const T*)(((const char*)data_) + offset_))[0];
-  }
-
-  /**
-   * Writes exactly *nbytes* into the local buffer by reading from the
-   * input buffer *buf*.
-   *
-   * @param buffer The buffer to read from.
-   * @param nbytes Number of bytes to write.
-   * @return Status.
-   */
-  Status write(const void* buffer, uint64_t nbytes);
-};
-
-}  // namespace sm
-}  // namespace tiledb
-
-#endif  // TILEDB_PREALLOCATED_BUFFER_H
